style(config): changed timeout typing to use float

+++ src/cb_events/config.py
@@ -27,7 +27,7 @@ class ClientConfig(BaseModel):
 
     model_config: ClassVar[ConfigDict] = ConfigDict(frozen=True)
 
-    timeout: int = Field(default=10, gt=0)
+    timeout: float = Field(default=10, gt=0)
     """Server long-poll timeout in seconds."""
 
     strict_validation: bool = False
